Refactored the code as per suggestions and also created new interface IJoinPredicate.
