+++ textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/dataflow/join/Join.java
@@ -1,22 +1,17 @@
 package edu.uci.ics.textdb.dataflow.join;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import edu.uci.ics.textdb.api.common.Attribute;
 import edu.uci.ics.textdb.api.common.FieldType;
-import edu.uci.ics.textdb.api.common.IField;
 import edu.uci.ics.textdb.api.common.ITuple;
 import edu.uci.ics.textdb.api.common.Schema;
 import edu.uci.ics.textdb.api.dataflow.IOperator;
 import edu.uci.ics.textdb.common.constants.SchemaConstants;
 import edu.uci.ics.textdb.common.exception.DataFlowException;
-import edu.uci.ics.textdb.common.field.DataTuple;
-import edu.uci.ics.textdb.common.field.ListField;
-import edu.uci.ics.textdb.common.field.Span;
-import edu.uci.ics.textdb.dataflow.common.JoinPredicate;
+import edu.uci.ics.textdb.dataflow.common.JoinDistancePredicate;
 
 
 /**
@@ -51,7 +46,7 @@ public class Join implements IOperator {
 
     private IOperator outerOperator;
     private IOperator innerOperator;
-    private JoinPredicate joinPredicate;
+    private JoinDistancePredicate joinPredicate;
     // To indicate if next result from outer operator has to be obtained.
     private boolean shouldIGetOuterOperatorNextTuple;
     private ITuple outerTuple = null;
@@ -75,13 +70,13 @@ public class Join implements IOperator {
      * @param joinPredicate
      *            is the predicate over which the join is made
      */
-    public Join(IOperator outerOperator, IOperator innerOperator, JoinPredicate joinPredicate) {
+    public Join(IOperator outerOperator, IOperator innerOperator, JoinDistancePredicate joinPredicate) {
         this.outerOperator = outerOperator;
         this.innerOperator = innerOperator;
         this.joinPredicate = joinPredicate;
     }
     
-    public Join(JoinPredicate joinPredicate) {
+    public Join(JoinDistancePredicate joinPredicate) {
         this.joinPredicate = joinPredicate;
     }
 
@@ -152,7 +147,7 @@ public class Join implements IOperator {
                 }
             }
             
-            nextTuple = joinTuples(outerTuple, innerTuple, joinPredicate);
+            nextTuple = joinPredicate.joinTuples(this, outerTuple, innerTuple);
         } while (nextTuple == null);
 
         return nextTuple;
@@ -172,116 +167,6 @@ public class Join implements IOperator {
         innerTupleList.clear();
     }
 
-    // compare if two tuples have the save value over a field
-    // return null if one of them doesn't have this field
-    private boolean compareField(ITuple innerTuple, ITuple outerTuple, String fieldName) {  
-        IField innerField = innerTuple.getField(fieldName);
-        IField outerField = outerTuple.getField(fieldName);
-        
-        if (innerField == null ||  outerField == null) {
-            return false;
-        }
-
-        return innerField.getValue().equals(outerField.getValue());
-    }
-
-    // Process the tuples to get a tuple with join result if predicate is
-    // satisfied.
-    private ITuple joinTuples(ITuple outerTuple, ITuple innerTuple, JoinPredicate joinPredicate) throws Exception {
-        List<Span> newJoinSpanList = new ArrayList<>();
-        
-        // We expect the values of all fields to be the same for innerTuple and outerTuple.
-        // We only checks ID field, and field to be joined, since they are crucial to join operator.
-        // For other fields, we use the value from innerTuple.
-
-        // check if the ID fields are the same
-        if (! compareField(innerTuple, outerTuple, joinPredicate.getIDAttribute().getFieldName())) {
-            return null;
-        }
-
-        // check if the fields to be joined are the same
-        if (! compareField(innerTuple, outerTuple, joinPredicate.getJoinAttribute().getFieldName())) {
-            return null;
-        }
-        
-        
-        // If either/both tuples have no span information, return null.
-        // Check using try/catch if both the tuples have span information.
-        // If not return null; so we can process next tuple.
-
-        IField spanFieldOfInnerTuple = null;
-        IField spanFieldOfOuterTuple = null;
-        try {
-            spanFieldOfInnerTuple = innerTuple.getField(SchemaConstants.SPAN_LIST);
-            spanFieldOfOuterTuple = outerTuple.getField(SchemaConstants.SPAN_LIST);
-        } catch (Exception e) {
-            return null;
-        }
-
-        List<Span> innerSpanList = null;
-        List<Span> outerSpanList = null;
-        // Check if both the fields obtained from the indexes are indeed of type
-        // ListField
-        if (spanFieldOfInnerTuple.getClass().equals(ListField.class)) {
-            innerSpanList = (List<Span>) spanFieldOfInnerTuple.getValue();
-        }
-        if (spanFieldOfOuterTuple.getClass().equals(ListField.class)) {
-            outerSpanList = (List<Span>) spanFieldOfOuterTuple.getValue();
-        }
-
-        Iterator<Span> outerSpanIter = outerSpanList.iterator();
-
-        // TODO Currently we are using two loops to go over two span lists.
-        // We can optimize it by sorting the spans based on the start position
-        // and then doing a "sort merge" of the two lists.
-        // (Also probably weed out the spans with fields that don't agree with
-        // the ones specified in the JoinPredicate during "sort merge"?)
-        while (outerSpanIter.hasNext()) {
-            Span outerSpan = outerSpanIter.next();
-            // Check if the field matches the filed over which we want to join.
-            // If not return null.
-            if (!outerSpan.getFieldName().equals(joinPredicate.getJoinAttribute().getFieldName())) {
-                continue;
-            }
-            Iterator<Span> innerSpanIter = innerSpanList.iterator();
-            while (innerSpanIter.hasNext()) {
-                Span innerSpan = innerSpanIter.next();
-                if (!innerSpan.getFieldName().equals(joinPredicate.getJoinAttribute().getFieldName())) {
-                    continue;
-                }
-                Integer threshold = joinPredicate.getThreshold();
-                if (Math.abs(outerSpan.getStart() - innerSpan.getStart()) <= threshold
-                        && Math.abs(outerSpan.getEnd() - innerSpan.getEnd()) <= threshold) {
-                    Integer newSpanStartIndex = Math.min(outerSpan.getStart(), innerSpan.getStart());
-                    Integer newSpanEndIndex = Math.max(outerSpan.getEnd(), innerSpan.getEnd());
-                    String fieldName = joinPredicate.getJoinAttribute().getFieldName();
-                    String fieldValue = (String) innerTuple.getField(fieldName).getValue();
-                    String newFieldValue = fieldValue.substring(newSpanStartIndex, newSpanEndIndex);
-                    String spanKey = outerSpan.getKey() + "_" + innerSpan.getKey();
-                    Span newSpan = new Span(fieldName, newSpanStartIndex, newSpanEndIndex, spanKey, newFieldValue);
-                    newJoinSpanList.add(newSpan);
-                }
-            }
-        }
-
-        if (newJoinSpanList.isEmpty()) {
-            return null;
-        }
-       
-        // create output fields based on innerTuple's value
-        List<IField> outputFields = 
-                outputAttrList.stream()
-                .filter(attr -> ! attr.equals(SchemaConstants.SPAN_LIST_ATTRIBUTE))
-                .map(attr -> attr.getFieldName())
-                .map(fieldName -> innerTuple.getField(fieldName))
-                .collect(Collectors.toList());
-        
-        outputFields.add(new ListField<>(newJoinSpanList));
-        
-        return new DataTuple(outputSchema, outputFields.stream().toArray(IField[]::new));
-    }
-    
-    
     /*
      * Create outputSchema, which is the intersection of innerOperator's schema and outerOperator's schema.
      * The attributes have to be exactly the same (name and type) to be intersected.
